requirements.txt
argparse
opencv-python
ultralytics
collections-extended
Pillow
configparser
requests

